moved arithmetic functions
